feat: backported 0.3.1 - Added `AlphaMode` API for better transparency control. - Implemented `BlockLightInfo` with adjustable intensity, radius, and falloff parameters. - Introduced a `ColorVariable` for dynamic light color management. - Added JSON configuration files for flexible runtime tuning of multi-threading, lights, and voxelized blocks (`default_config.json`). - Created unit tests for block building, registry parity, and voxel packing to ensure 3.1.0 compatibility. - Added RenderDoc analysis script for debugging graphics artifacts in frames. - Updated mixins and adapters for improved shader integration and resource management. - Removed `BlockModelRendererMixin` and unused photonic patches for cleaner implementation.

+++ src/at/redi2go/photonic/client/rendering/util/BufferUtils.java
@@ -54,11 +54,11 @@ public class BufferUtils {
       return ar;
    }
 
-   public static byte[] packUnorm4x8(float... data) {
-      byte[] bytes = new byte[data.length];
+   public static int[] packUnorm4x8(float... data) {
+      int[] bytes = new int[data.length];
 
       for (int i = 0; i < data.length; i++) {
-         bytes[i] = (byte)Math.round(Math.clamp((double)data[i], 0.0, 1.0) * 255.0);
+         bytes[i] = (int)Math.round(Math.clamp((double)data[i], 0.0, 1.0) * 255.0);
       }
 
       return bytes;
